Add files via upload
Comments are given in each code regarding contribution of every members. Mine is under the name of Leong Kai Ler and my contributions are mainly on design and planning the structure of the code, coding the graphs and charts and designing the maps. Aesthetic modifications are done by me on the charts, graphs and maps.